+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `auth` command to generate PlotNFT login links.
+
 ## [1.3.0] - 2023-10-25
 
 ### Added